// src/app/modules/admin/g-admin/vendor/vendor.types.ts
export interface BranchPagination {
    length: number;
    size: number;
    page: number;
    lastPage: number;
    startIndex: number;
    endIndex: number;
    pagination: {
        page: number,
        pageSize: number,
        pageCount: number,
        total: number
    }
}
// ...
export interface AssetType {

    id: string;
    name: string,
    asset_sizes: [
        {
            id: number,
            name: string,
            size: string,
        }
    ]
}
// ...
export interface DataVendor {
    data: [
        {
            id: number,
            name: string,
            contact: string,
            email: string,
            phone: number,
            adress: string,
            status: number,
            create_by: number,
            update_by: string,
            created_at: Date,
            updated_at: Date,
            No: number
        }
    ]
}